requirements.txt
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
anthropic==0.7.8
python-dotenv==1.0.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2
aiofiles==23.2.1
openpyxl==3.1.2

